Two more nodes
'Compare' extends the 'Greater Than' and 'Less Than' from the Math node. It has also 'Greater Than or Equal To',  'Less Than or Equal To',  'Equal To' and 'Not Equal To'.

'Switch Float' it's a simple switcher. If 'switch' is 0 it outputs 'value1', if 1 it outputs 'value2'.

Both can be found in the 'Converter' category.